requirements.txt
asgiref==3.8.1
babel==2.16.0
backports.zoneinfo==0.2.1
Django==4.2.15
django-admin-confirm==1.0.0
django-ckeditor-5==0.2.13
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django-seed==0.3.1
django-widget-tweaks==1.5.0
Faker==28.4.1
module-name==0.6.0
mysqlclient==2.2.4
phonenumbers==8.13.43
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
